Add files via upload
The goal of the research is to help the auditors by building a classification model that can predict the fraudulent firm on the basis of the present and historical risk factors. This is a Machine Learning Classification problem.
Models used- Logistic Regression, Decision Tree, Random Forest, AdaBoost, XGBoost, SVM, Naive-Bayes, K-Neighbours Classifier.